+++ src/client/transport.rs
@@ -179,6 +179,7 @@ mod tests {
             coin_spends: vec![],
             required_signatures: vec![],
             summary: TransactionSummary {
+                received: vec![],
                 outputs: vec![xch_output()],
                 fee: Amount(1),
             },
@@ -257,6 +258,7 @@ mod tests {
             coin_spends: vec![],
             required_signatures: vec![],
             summary: TransactionSummary {
+                received: vec![],
                 outputs: vec![],
                 fee: Amount(0),
             },